# lib/fs: Optimize UnicodeLowercase