Most notably, it now detects all-lowercase files and returns these as-is. The tests have been expanded with two cases and are now used as a benchmark (admittedly a rather trivial one).

Benchmark results on Linux/amd64 (note that WalkCaseFakeFS100k creates all-lowercase filenames):

    name                           old time/op         new time/op         delta
    WalkCaseFakeFS100k/rawfs-8             388ms ± 1%          312ms ± 3%   -19.64%  (p=0.000 n=10+8)
    WalkCaseFakeFS100k/casefs-8            718ms ± 1%          587ms ± 1%   -18.30%  (p=0.000 n=10+10)
    UnicodeLowercaseMaybeChange-8         5.45µs ± 5%         4.57µs ± 1%   -16.20%  (p=0.000 n=10+10)
    UnicodeLowercaseNoChange-8            5.33µs ± 4%         3.09µs ± 1%   -42.01%  (p=0.000 n=10+10)

    name                           old B/entry         new B/entry         delta
    WalkCaseFakeFS100k/rawfs-8               590 ± 0%            530 ± 0%   -10.17%  (p=0.000 n=10+10)
    WalkCaseFakeFS100k/casefs-8              869 ± 0%            760 ± 0%   -12.54%  (p=0.000 n=10+10)
    
    name                           old allocs/entry    new allocs/entry    delta
    WalkCaseFakeFS100k/rawfs-8              15.7 ± 0%           12.1 ± 0%   -22.93%  (p=0.000 n=10+10)
    WalkCaseFakeFS100k/casefs-8             26.1 ± 0%           20.1 ± 0%   -22.99%  (p=0.000 n=10+10)
    
    name                           old ns/entry        new ns/entry        delta
    WalkCaseFakeFS100k/rawfs-8             2.57k ± 1%          2.06k ± 3%   -19.65%  (p=0.000 n=10+8)
    WalkCaseFakeFS100k/casefs-8            4.75k ± 1%          3.86k ± 1%   -18.63%  (p=0.000 n=10+10)
    
    name                           old alloc/op        new alloc/op        delta
    WalkCaseFakeFS100k/rawfs-8            89.2MB ± 0%         80.2MB ± 0%   -10.08%  (p=0.000 n=9+10)
    WalkCaseFakeFS100k/casefs-8            131MB ± 0%          115MB ± 0%   -12.54%  (p=0.000 n=9+10)
    UnicodeLowercaseMaybeChange-8           320B ± 0%           192B ± 0%   -40.00%  (p=0.000 n=10+10)
    UnicodeLowercaseNoChange-8              320B ± 0%             0B       -100.00%  (p=0.000 n=10+10)
    
    name                           old allocs/op       new allocs/op       delta
    WalkCaseFakeFS100k/rawfs-8             2.38M ± 0%          1.82M ± 0%   -23.44%  (p=0.002 n=8+10)
    WalkCaseFakeFS100k/casefs-8            3.95M ± 0%          3.04M ± 0%   -23.10%  (p=0.000 n=10+9)
    UnicodeLowercaseMaybeChange-8           16.0 ± 0%           10.0 ± 0%   -37.50%  (p=0.000 n=10+10)
    UnicodeLowercaseNoChange-8              16.0 ± 0%            0.0       -100.00%  (p=0.000 n=10+10)