@greatroar 2020-09-10
I must confess that I don't know what's causing

    UnicodeLowercaseMaybeChange-8           320B ± 0%           384B ± 0%   +20.00%  (p=0.000 n=10+10)

Overallocating a few bytes (say, 4 or 10) can decrease the alloc/op to 208B, but I don't understand why. The encoded output is never longer than the encoded input in the benchmark.