@greatroar 2020-09-10
Found it, see updated benchmark results and comment in the code.