@calmh 2020-09-10
So this is neat, though I think the benchmarks aren't useful here. The improvement is equally large for the `raw` case that doesn't use `casefs`, since the underlying fakefs itself uses UnicodeLowercase a lot to implement its fake case insensitivity.