@greatroar 2020-09-10
> The improvement is equally large for the raw case that doesn't use casefs, since the underlying fakefs itself uses UnicodeLowercase a lot to implement its fake case insensitivity.

I'm not sure I follow completely. Should I remove the rawfs benchmark results from the commit message? The casefs ones? Both?